Fixed download progress announcement.
Simple progress description now shows download speed as well.
That was useful in diagnosing issues.
I don't plan on adding logic for a progress bar, that seems
like a lot of work for a really limited improvement in usability.
My next goal is to allow the script to try different mirrors
on a curl failure, then I think this script is complete.